requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
contourpy==1.2.0
cycler==0.12.1
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.0
fonttools==4.49.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
importlib_resources==6.1.2
iniconfig==2.0.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
priority==2.0.0
pydantic==2.6.2
pydantic_core==2.16.3
pymongo==4.6.2
pyparsing==3.1.1
pytest==8.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
taskgroup==0.0.0a4
tomli==2.0.1
typing_extensions==4.10.0
wsproto==1.2.0
zipp==3.17.0
